Changed Exercise 2 to use reset:true in fetch and updated discussion accordingly.
Technically, the code still works without the change, but only because adds
are triggered instead (renderBook gets called instead of render when the
fetch completes). Note: the render in the initialize() could be removed
at the same time the reset listener is added since the collection will
always be initially empty at this point in the example.